Get the game sources largely compiled(assembled)
...but they do not link.

The game uses base-relative addressing of the __MERGED section via a5.
Thats why in System we provide 'basereg __MERGED,a5'
But for some reason vasm won't have it and claims illegal relocations.

When linking with gcc, we'd need to provide the _LinkerDB symbol that
is being referenced. vlink should be able to do this on its own.