Work on getting the base-relative adressing working
Breathless produces 3 sections
1. CODE,bss for code and constant tables
2. __MERGED,bss for base-relative addressed runtime data
3. TABLES,bss for large tables that don't fit into __MERGED

Provide '-sd -sdreg=5' to casm to enable base-relative adressing via a5.
Fix all occurrences of 'lea	ObjectsPunList-4(a5),aX' (which vasm ddoesn't like)
to refer to 'ObjectsPunListMinus4' which is placed directly in front of
ObjectsPunList, thus should achieve the same effect.

The Linker is supposed to place the Symbol '_LinkerDB' into the data segment
as base for the bas-relative addressing. Right now, this symbol will end up
in the TABLES section. That needs to be fixed somehow.

Additionally, use '.text' as the name for the code section such that
the produced object files are compatible with Bebbo's GCC tools and
we can link against GCC's libamiga.a without creating two different
code sections (.text and CODE)

Provide TMapMain as entry point to the linker.