Fix Enforcer hits
The original code attempted to place "ChunkyPointer" exactly at the
"start" of the BSS segment to allow for a 0-offset acccess.

But in our case, ChunkyPointer sits right at the start of the BSS segment at -0x7FFE.
The BSS segment is not even 32kb in size, so forcing ChunkyPointer to be at 0(a5)
instead of -7FFE(a5) is causing access to invalid memory.

I could attempt to just artificially enlarge the __MERGED segment to cover the 0(a5)
address... but for now I'll play it safe.